pyproject.toml
[project]
name = "thesis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "asyncio>=3.4.3",
    "asyncpg>=0.30.0",
    "black>=25.1.0",
    "chainlit>=2.6.2",
    "ipykernel>=6.29.5",
    "llama-index>=0.12.50",
    "loguru>=0.7.3",
    "milvus-lite>=2.5.1",
    "mypy>=1.17.0",
    "openai>=1.95.0",
    "openrouter>=1.0",
    "pandas>=2.3.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pymilvus>=2.5.13",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "python-dotenv>=1.1.1",
    "redis>=6.2.0",
    "ruff>=0.12.4",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.41",
    "sqlmodel>=0.0.24",
]
